minor MusyX fixes
Former-commit-id: ce047bc95cd9421978c6e194e5a02e3e0f8425b2

+++ src/musyx/runtime/synth_vsamples.c
@@ -301,11 +301,11 @@ void sndVirtualSampleARAMUpdate(unsigned short instID, void* base, unsigned long
 
     if (len1 != 0) {
       hwFlushStream(base, off1, len1, vs.streamBuffer[i].hwId, vsARAMDMACallback,
-                    &vs.streamBuffer[i]);
+                    (u32)&vs.streamBuffer[i]);
     }
     if (len2 != 0) {
       hwFlushStream(base, off2, len2, vs.streamBuffer[i].hwId, vsARAMDMACallback,
-                    &vs.streamBuffer[i]);
+                    (u32)&vs.streamBuffer[i]);
     }
 
     if (vs.streamBuffer[i].smpType == 5) {
